Move buttons to a separate file; fix disabled state
`buttons.tsx` to store most of the buttons, so logic will be more readable in `Controls.tsx`.
Buttons which can require a comment use handler wrapper.
`disabled` is added to custom buttons as well to block them during action handling as all other buttons.
